<REFACTOR> sperate application logic from host
