+++ src/main/java/com/instream/tenant/domain/host/config/HostRouterConfig.java
@@ -1,21 +1,12 @@
 package com.instream.tenant.domain.host.config;
 
-import com.google.common.reflect.TypeToken;
-import com.instream.tenant.domain.application.domain.dto.ApplicationDto;
-import com.instream.tenant.domain.application.domain.request.ApplicationCreateRequest;
-import com.instream.tenant.domain.application.domain.request.ApplicationSearchPaginationOptionRequest;
-import com.instream.tenant.domain.application.domain.response.ApplicationCreateResponse;
-import com.instream.tenant.domain.common.domain.dto.CollectionDto;
-import com.instream.tenant.domain.common.domain.dto.PaginationDto;
 import com.instream.tenant.domain.host.domain.dto.TenantDto;
 import com.instream.tenant.domain.host.domain.request.TenantCreateRequest;
 import com.instream.tenant.domain.host.domain.request.TenantSignInRequest;
 import com.instream.tenant.domain.host.handler.HostHandler;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import org.springdoc.core.fn.builders.apiresponse.Builder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.web.reactive.function.server.RequestPredicates;
 import org.springframework.web.reactive.function.server.RouterFunction;
 import org.springframework.web.reactive.function.server.ServerResponse;
@@ -31,14 +22,12 @@ import static org.springdoc.webflux.core.fn.SpringdocRouteBuilder.route;
 @Configuration
 public class HostRouterConfig {
     @Bean
-    public RouterFunction<ServerResponse> v1Routes(HostHandler hostHandler) {
+    public RouterFunction<ServerResponse> v1HostRoutes(HostHandler hostHandler) {
         return route().nest(RequestPredicates.path("/v1/hosts"),
                 builder -> {
                     builder.add(getUserInfoFunctionSupplier(hostHandler));
                     builder.add(signInFunctionSupplier(hostHandler));
                     builder.add(signUpFunctionSupplier(hostHandler));
-                    builder.add(searchApplication(hostHandler));
-                    builder.add(createApplication(hostHandler));
                 },
                 ops -> ops.operationId("123")
         ).build();
@@ -81,35 +70,6 @@ public class HostRouterConfig {
                         ops -> ops.operationId("123")
                                 .requestBody(requestBodyBuilder().implementation(TenantCreateRequest.class))
                                 .response(responseBuilder().implementation(TenantDto.class))
-                )
-                .build();
-    }
-
-    private RouterFunction<ServerResponse> searchApplication(HostHandler hostHandler) {
-        return route()
-                .GET(
-                        "/{hostId}/applications",
-                        hostHandler::searchApplication,
-                        ops -> ops.operationId(String.format("pagination_%s", ApplicationDto.class.getSimpleName()))
-                                .parameter(parameterBuilder().name("hostId").in(ParameterIn.PATH).required(true).example("80bd6328-76a7-11ee-b720-0242ac130003"))
-                                .parameter(parameterBuilder().in(ParameterIn.QUERY).name("option").implementation(ApplicationSearchPaginationOptionRequest.class))
-                )
-                .build();
-    }
-
-    private RouterFunction<ServerResponse> createApplication(HostHandler hostHandler) {
-        return route()
-                .POST(
-                        "/{hostId}/applications",
-                        hostHandler::createApplication,
-                        ops -> ops.operationId("123")
-                                .parameter(parameterBuilder()
-                                        .name("hostId")
-                                        .in(ParameterIn.PATH)
-                                        .required(true)
-                                        .example("80bd6328-76a7-11ee-b720-0242ac130003"))
-                                .requestBody(requestBodyBuilder().implementation(ApplicationCreateRequest.class))
-                                .response(responseBuilder().implementation(ApplicationCreateResponse.class))
                 )
                 .build();
     }
